added controller superclass.
