+++ app/controllers/Application.java
@@ -7,24 +7,7 @@ import play.mvc.*;
 import controllers.Secure;
 import models.*;
 
-@With(Secure.class)
-public class Application extends Controller {
-
-  @Before
-  static void setConnectedUser() {
-    if (Security.isConnected()) {
-      renderArgs.put("currentUser", user());
-    }
-  }
-
-  @Before
-  static void addDefaults() {
-  }
-
-  public static User user() {
-    assert Secure.Security.connected() != null;
-    return User.getUser(Secure.Security.connected());
-  }
+public class Application extends OBController {
 
   public static void about(Long id) {
     User user = id == null ? user() : (User) User.findById(id);
