+++ app/controllers/Statuses.java
@@ -9,22 +9,22 @@ import models.*;
 import play.data.validation.*;
 
 @With(Secure.class)
-public class Statuses extends Controller {
-    
+public class Statuses extends OBController {
+
   public static void show(Long id){
     Status status = Status.findById(id);
     User author = User.findById(id);
     render(status, author);
   }
-  
+
   public static void postStatus(
-    @Required(message="A message is required") String content) 
+    @Required(message="A message is required") String content)
   {
-    
+
     Status status = new Status(Application.user(), content).save();
     Application.news(Application.user().id);
   }
-  
+
   public List<Status> returnAll(){
     return Status.findAll();
   }
